Split out WebClient Cache module
